requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-socketio
cryptography
pynacl
python-dotenv
google-generativeai
requests